chore: split raster module into files
